+++ src/subtitles/VobSubFileRipper.cpp
@@ -555,7 +555,7 @@ bool CVobSubFileRipper::Create()
 
 	if(m_rd.fResetTime)
 	{
-		for(int i = 0; i < angle.GetCount() && ((angle[i].vob<<16)|angle[i].cell) != m_rd.selvcs[0]; i++)
+		for(size_t i = 0; i < angle.GetCount() && ((angle[i].vob<<16)|angle[i].cell) != m_rd.selvcs[0]; i++)
 			tStart += angle[i].tTime;
 
 		Log(LOG_INFO, _T("Counting timestamps from %I64dms (v%02dc%02d)"), 
@@ -564,7 +564,7 @@ bool CVobSubFileRipper::Create()
 
 	CAtlMap<DWORD, int> selvcmap;
 	selvcmap.RemoveAll();
-	for(int i = 0; i < m_rd.selvcs.GetCount(); i++)
+	for(size_t i = 0; i < m_rd.selvcs.GetCount(); i++)
 		selvcmap[m_rd.selvcs[i]] = 90000;
 
 	CAtlArray<vcchunk> chunks, foundchunks, loadedchunks;
@@ -573,7 +573,7 @@ bool CVobSubFileRipper::Create()
 	{
 		Log(LOG_INFO, _T("Indexing mode: DVD"));
 
-		for(int i = 0; i < angle.GetCount(); i++)
+		for(size_t i = 0; i < angle.GetCount(); i++)
 		{
 			DWORD vc = (angle[i].vob<<16)|angle[i].cell;
 			if(!selvcmap.Lookup(vc))
@@ -590,7 +590,7 @@ bool CVobSubFileRipper::Create()
 	{
 		Log(LOG_INFO, _T("Indexing mode: File"));
 
-		for(int i = 0; i < loadedchunks.GetCount(); i++)
+		for(size_t i = 0; i < loadedchunks.GetCount(); i++)
 		{
 			DWORD vcid = loadedchunks[i].vc;
 			if(!selvcmap.Lookup(vcid))
@@ -611,10 +611,10 @@ bool CVobSubFileRipper::Create()
 	}
 
 	__int64 sizedone = 0, sizetotal = 0;
-	for(int i = 0; i < chunks.GetCount(); i++)
+	for(size_t i = 0; i < chunks.GetCount(); i++)
 		sizetotal += chunks[i].end - chunks[i].start;
 
-	for(int i = 0; !m_fBreakThread && i < chunks.GetCount(); i++)
+	for(size_t i = 0; !m_fBreakThread && i < chunks.GetCount(); i++)
 	{
 		__int64 curpos = chunks[i].start, endpos = chunks[i].end;
 
@@ -719,7 +719,7 @@ bool CVobSubFileRipper::Create()
 
 				tOffset = tTotal = 0;
 
-				for(int i = 0; i < angle.GetCount(); i++)
+				for(size_t i = 0; i < angle.GetCount(); i++)
 				{
 					if(angle[i].vob == vob && angle[i].cell == cell)
 					{
@@ -1033,7 +1033,7 @@ STDMETHODIMP CVobSubFileRipper::LoadParamFile(CString fn)
 						s++;
 						if(vob != 0 && cell == 0)
 						{
-							for(int i = 0; i < angle.GetCount(); i++)
+							for(size_t i = 0; i < angle.GetCount(); i++)
 							{
 								if(angle[i].vob == vob)
 									m_rd.selvcs.Add((angle[i].vob<<16)|angle[i].cell);
@@ -1048,7 +1048,7 @@ STDMETHODIMP CVobSubFileRipper::LoadParamFile(CString fn)
 						s++;
 						cell = _tcstol(s, &s, 10);
 
-						for(int i = 0; i < angle.GetCount(); i++)
+						for(size_t i = 0; i < angle.GetCount(); i++)
 						{
 							if(angle[i].vob == vob && angle[i].cell == cell)
 							{
@@ -1065,7 +1065,7 @@ STDMETHODIMP CVobSubFileRipper::LoadParamFile(CString fn)
 			}
 			else
 			{
-				for(int i = 0; i < angle.GetCount(); i++)
+				for(size_t i = 0; i < angle.GetCount(); i++)
 					m_rd.selvcs.Add((angle[i].vob<<16)|angle[i].cell);
 			}
